all /components/search/svg/ files
