+++ website/components/search/svg/CancelIcon.tsx
@@ -1,4 +1,4 @@
-export default function RightArrow({}) {
+export default function CancelIcon() {
     return (
         <svg
             xmlns="http://www.w3.org/2000/svg"
@@ -9,9 +9,10 @@ export default function RightArrow({}) {
             stroke="currentColor"
             strokeWidth="2"
             strokeLinecap="round"
-            strokeLinejoin="round"
-        >
-            <polyline points="9 18 15 12 9 6"></polyline>
+            strokeLinejoin="round">
+            <circle cx="12" cy="12" r="10"></circle>
+            <line x1="15" y1="9" x2="9" y2="15"></line>
+            <line x1="9" y1="9" x2="15" y2="15"></line>
         </svg>
     );
 }
