+++ website/components/search/svg/DoubleLeftArrowsIcon.tsx
@@ -1,4 +1,4 @@
-export default function LeftArrow({}) {
+export default function DoubleLeftArrowsIcon() {
     return (
         <svg
             xmlns="http://www.w3.org/2000/svg"
@@ -9,9 +9,9 @@ export default function LeftArrow({}) {
             stroke="currentColor"
             strokeWidth="2"
             strokeLinecap="round"
-            strokeLinejoin="round"
-        >
-            <polyline points="15 18 9 12 15 6"></polyline>
+            strokeLinejoin="round">
+            <polyline points="11 17 6 12 11 7"></polyline>
+            <polyline points="18 17 13 12 18 7"></polyline>
         </svg>
     );
 }
