+++ website/components/search/svg/DoubleRightArrowsIcon.tsx
@@ -1,4 +1,4 @@
-export default function DoubleLeftArrowsIcon({}) {
+export default function DoubleRightArrowIcons() {
     return (
         <svg
             xmlns="http://www.w3.org/2000/svg"
@@ -9,10 +9,9 @@ export default function DoubleLeftArrowsIcon({}) {
             stroke="currentColor"
             strokeWidth="2"
             strokeLinecap="round"
-            strokeLinejoin="round"
-        >
-            <polyline points="11 17 6 12 11 7"></polyline>
-            <polyline points="18 17 13 12 18 7"></polyline>
+            strokeLinejoin="round">
+            <polyline points="13 17 18 12 13 7"></polyline>
+            <polyline points="6 17 11 12 6 7"></polyline>
         </svg>
     );
 }
