+++ website/components/search/svg/DownArrow.tsx
@@ -1,4 +1,4 @@
-export default function DoubleRightArrowIcons({}) {
+export default function DownArrow() {
     return (
         <svg
             xmlns="http://www.w3.org/2000/svg"
@@ -9,10 +9,8 @@ export default function DoubleRightArrowIcons({}) {
             stroke="currentColor"
             strokeWidth="2"
             strokeLinecap="round"
-            strokeLinejoin="round"
-        >
-            <polyline points="13 17 18 12 13 7"></polyline>
-            <polyline points="6 17 11 12 6 7"></polyline>
+            strokeLinejoin="round">
+            <polyline points="6 9 12 15 18 9"></polyline>
         </svg>
     );
 }
