+++ website/components/search/svg/LeftArrow.tsx
@@ -1,4 +1,4 @@
-export default function CancelIcon() {
+export default function LeftArrow() {
     return (
         <svg
             xmlns="http://www.w3.org/2000/svg"
@@ -9,11 +9,8 @@ export default function CancelIcon() {
             stroke="currentColor"
             strokeWidth="2"
             strokeLinecap="round"
-            strokeLinejoin="round"
-        >
-            <circle cx="12" cy="12" r="10"></circle>
-            <line x1="15" y1="9" x2="9" y2="15"></line>
-            <line x1="9" y1="9" x2="15" y2="15"></line>
+            strokeLinejoin="round">
+            <polyline points="15 18 9 12 15 6"></polyline>
         </svg>
     );
 }
