+++ website/components/search/svg/SearchBarIcon.tsx
@@ -1,4 +1,4 @@
-export default function DownArrow({}) {
+export default function SearchBarIcon() {
     return (
         <svg
             xmlns="http://www.w3.org/2000/svg"
@@ -9,9 +9,9 @@ export default function DownArrow({}) {
             stroke="currentColor"
             strokeWidth="2"
             strokeLinecap="round"
-            strokeLinejoin="round"
-        >
-            <polyline points="6 9 12 15 18 9"></polyline>
+            strokeLinejoin="round">
+            <circle cx="11" cy="11" r="8"></circle>
+            <line x1="21" y1="21" x2="16.65" y2="16.65"></line>
         </svg>
     );
 }
